requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
PyYAML==6.0.1
